moved project to root directory in repo
