--- krowd/src/main/java/com/revature/krowd/KrowdApplication.java
@@ -1,14 +1,0 @@
-package com.revature.krowd;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class KrowdApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(KrowdApplication.class, args);
-	}
-
-}
-
