--- krowd/src/main/java/com/revature/krowd/bean/Admin.java
@@ -1,121 +1,0 @@
-package com.revature.krowd.bean;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import javax.validation.constraints.NotNull;
-
-@Entity
-@Table(name = "admin")
-public class Admin implements Serializable {
-	
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = -8360921354766267159L;
-	
-	public Admin() {
-		super();
-	}
-
-	public Admin(int id, @NotNull String username, @NotNull String email, @NotNull String firstname,
-			@NotNull String lastname) {
-		super();
-		this.id = id;
-		this.username = username;
-		this.email = email;
-		this.firstname = firstname;
-		this.lastname = lastname;
-	}
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
-	
-	@Column
-	@NotNull
-	private String username;
-	
-	@Column
-	@NotNull
-	private String email;
-	
-	@Column
-	@NotNull
-	private String firstname;
-	
-	@Column
-	@NotNull
-	private String lastname;
-
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-
-	public String getUsername() {
-		return username;
-	}
-
-	public void setUsername(String username) {
-		this.username = username;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getFirstname() {
-		return firstname;
-	}
-
-	public void setFirstname(String firstname) {
-		this.firstname = firstname;
-	}
-
-	public String getLastname() {
-		return lastname;
-	}
-
-	public void setLastname(String lastname) {
-		this.lastname = lastname;
-	}
-
-	@Override
-	public String toString() {
-		return "Admin [id=" + id + ", username=" + username + ", email=" + email + ", firstname=" + firstname
-				+ ", lastname=" + lastname + "]";
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(email, firstname, id, lastname, username);
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		Admin other = (Admin) obj;
-		return Objects.equals(email, other.email) && Objects.equals(firstname, other.firstname) && id == other.id
-				&& Objects.equals(lastname, other.lastname) && Objects.equals(username, other.username);
-	}
-
-}
