--- krowd/src/main/java/com/revature/krowd/bean/EventCategory.java
@@ -1,79 +1,0 @@
-package com.revature.krowd.bean;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import javax.validation.constraints.NotNull;
-
-@Entity
-@Table(name = "event_category")
-public class EventCategory implements Serializable {
-	
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = -424087293988598016L;
-
-	public EventCategory() {
-		super();
-	}
-
-	public EventCategory(int id, @NotNull String name) {
-		super();
-		this.id = id;
-		this.name = name;
-	}
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
-	
-	@Column
-	@NotNull
-	private String name;
-
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	@Override
-	public String toString() {
-		return "EventCategory [id=" + id + ", name=" + name + "]";
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(id, name);
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		EventCategory other = (EventCategory) obj;
-		return id == other.id && Objects.equals(name, other.name);
-	}
-
-}
