--- krowd/src/test/java/com/revature/krowd/KrowdApplicationTests.java
@@ -1,17 +1,0 @@
-package com.revature.krowd;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class KrowdApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
-
