+++ src/main/java/com/revature/krowdboot/KrowdBootApplication.java
@@ -1,0 +1,14 @@
+package com.revature.krowdboot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class KrowdBootApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(KrowdBootApplication.class, args);
+	}
+
+}
+
