+++ src/main/java/com/revature/krowdboot/repository/AdminRepository.java
@@ -1,0 +1,11 @@
+package com.revature.krowdboot.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.revature.krowdboot.model.Admin;
+
+@Repository
+public interface AdminRepository extends JpaRepository<Admin, Integer>{
+
+}
