+++ src/main/java/com/revature/krowdboot/repository/EventCategory.java
@@ -1,0 +1,9 @@
+package com.revature.krowdboot.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface EventCategory extends JpaRepository<EventCategory, Integer> {
+
+}
