+++ src/main/java/com/revature/krowdboot/repository/EventRepository.java
@@ -1,0 +1,11 @@
+package com.revature.krowdboot.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.revature.krowdboot.model.Event;
+
+@Repository
+public interface EventRepository extends JpaRepository<Event, Integer> {
+
+}
