+++ src/main/resources/application.properties
@@ -1,0 +1,21 @@
+# H2
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2
+
+
+# Datasource
+spring.datasource.url=jdbc:h2:~/test
+spring.datasource.username=sa
+spring.datasource.password=
+spring.datasource.driver-class-name=org.h2.Driver
+
+# Hibernate
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.show-sql=true
+
+
+# Port
+server.port=8085
+
+# Jackson
+#spring.jackson.serialization.fail-on-empty-beans=false
