+++ src/test/java/com/revature/krowdboot/KrowdBootApplicationTests.java
@@ -1,0 +1,17 @@
+package com.revature.krowdboot;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class KrowdBootApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
+
